refactor: alterando estrutura de pastas
